Move astar/simplify to utilities, add route option to node-drag
