+++ src/addons/node-drag.js
@@ -1,12 +1,17 @@
 import * as d3 from 'd3';
 import { flatten } from '../utils';
 import { translate } from '../utils/svg-util';
+import { getAStarPath } from '../utils/a-star';
+// import { simplifyPath } from '../utils/simplify';
+
 
 const nodeDrag = (G) => {
+  const edgeTracker = new Map();
+
   /**
    * Enable node dragging, this will recalculate edge end points as well
    */
-  const enableDrag = () => {
+  const enableDrag = (useAStarRouting) => {
     const chart = G.chart;
     const data = flatten(G.layout);
 
@@ -49,12 +54,15 @@ const nodeDrag = (G) => {
             p.x += dx;
             p.y += dy;
           });
+          edgeTracker.set(edge.id, 1);
         } else if (draggedIds.includes(source)) {
           edge.points[0].x += dx;
           edge.points[0].y += dy;
+          edgeTracker.set(edge.id, 1);
         } else if (draggedIds.includes(target)) {
           edge.points[edge.points.length - 1].x += dx;
           edge.points[edge.points.length - 1].y += dy;
+          edgeTracker.set(edge.id, 1);
         }
       });
 
@@ -62,7 +70,35 @@ const nodeDrag = (G) => {
       G.updateEdgePoints();
     }
 
-    function dragEnd() {}
+    function dragEnd() {
+      const collisionFn = (p) => {
+        const buffer = 10;
+        for (let i = 0; i < data.nodes.length; i++) {
+          const node = data.nodes[i];
+          if (node.nodes) return false;
+          if (p.x >= node.x - buffer && p.x <= node.x + node.width + buffer) {
+            if (p.y >= node.y - buffer && p.y <= node.y + node.height + buffer) {
+              return true;
+            }
+          }
+        }
+        return false;
+      };
+
+      if (useAStarRouting === true) {
+        data.edges.forEach(e => {
+          const points = e.points;
+          const start = points[0];
+          const end = points[points.length - 1];
+          if (e.source === e.target) return;
+          if (edgeTracker.has(e.id) === false) return;
+
+          e.points = getAStarPath(start, end, collisionFn, { w: 20, h: 20 });
+        });
+        G.updateEdgePoints();
+      }
+      edgeTracker.clear();
+    }
 
     // FIXME: Need to disable current listeners first before assigning new ones?
     const nodeDrag = d3.drag()
